+++ provider/resource_repository_test.go
@@ -17,8 +17,55 @@ func TestAccRepositoryBasic(t *testing.T) {
 		CheckDestroy:      testAccCheckRepositoryDestroy,
 		Steps: []resource.TestStep{
 			{
-				Config: basicRepositoryWithDescription("this is a testrepo"),
+				Config: basicRepositoryWithDescription("this is a testrepo", ""),
 				Check:  resource.TestCheckResourceAttr("scm_repository.testrepo", "id", "scmadmin/testrepo"),
+			},
+		},
+	})
+}
+
+func TestAccRepositoryUpdates(t *testing.T) {
+	resource.Test(t, resource.TestCase{
+		PreCheck:          func() { testAccPreCheck(t) },
+		ProviderFactories: testAccProviders,
+		CheckDestroy:      testAccCheckRepositoryDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: basicRepositoryWithDescription("this is a testrepo", ""),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttr("scm_repository.testrepo", "id", "scmadmin/testrepo"),
+					resource.TestCheckResourceAttr("scm_repository.testrepo", "last_modified", "")),
+			},
+			{
+				Config: basicRepositoryWithDescription("this is new description", ""),
+				Check:  resource.TestCheckResourceAttr("scm_repository.testrepo", "description", "this is new description"),
+			},
+			{
+				Config: basicRepositoryWithDescription("this is new description2", ""),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttr("scm_repository.testrepo", "description", "this is new description2"),
+					resource.TestCheckResourceAttrSet("scm_repository.testrepo", "last_modified")),
+			},
+		},
+	})
+}
+
+func TestAccRepositoryImport(t *testing.T) {
+	resource.Test(t, resource.TestCase{
+		PreCheck:          func() { testAccPreCheck(t) },
+		ProviderFactories: testAccProviders,
+		CheckDestroy:      testAccCheckRepositoryDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: basicRepositoryWithDescription("this is a testrepo", ""),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttr("scm_repository.testrepo", "id", "scmadmin/testrepo"),
+					resource.TestCheckResourceAttr("scm_repository.testrepo", "last_modified", "")),
+			},
+			{
+				Config: basicRepositoryWithDescription("this is new description", "import_url = \"https://github.com/cloudogu/spring-petclinic\""),
+				// For now there is no real check whether the import was successful
+				Check: resource.TestCheckResourceAttr("scm_repository.testrepo", "description", "this is new description"),
 			},
 		},
 	})
@@ -43,7 +90,7 @@ func testAccCheckRepositoryDestroy(s *terraform.State) error {
 	return nil
 }
 
-func basicRepositoryWithDescription(description string) string {
+func basicRepositoryWithDescription(description string, additionalFields string) string {
 	return fmt.Sprintf(`
 resource "scm_repository" "testrepo" {
   namespace = "scmadmin"
@@ -51,6 +98,7 @@ resource "scm_repository" "testrepo" {
   type = "git"
   description = "%s"
   contact = "scmadmin@test.test"
+  %s
 }
-`, description)
+`, description, additionalFields)
 }