+++ CHANGELOG.md
@@ -7,5 +7,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-* initial provider
-* scm_repository resource
+* initial provider [details](https://github.com/cloudogu/terraform-provider-scm/tree/95c10656978e96f7d0613933c0e35b7fb5be606a#provider-configuration)
+* scm_repository resource [details](https://github.com/cloudogu/terraform-provider-scm/tree/95c10656978e96f7d0613933c0e35b7fb5be606a#scm_repository)